+++ libraries/DataFlash/DataFlash_APM2.h
@@ -24,7 +24,7 @@ class DataFlash_APM2 : public DataFlash_Class
 	void CS_active();
 	void PageErase (uint16_t PageAdr);
 	void BlockErase (uint16_t BlockAdr);
-	void ChipErase(void (*delay_cb)(unsigned long));
+	void ChipErase(void (*delay_cb)(uint32_t));
 
   public:
 	DataFlash_APM2(); // Constructor
